Fixed bug in which google translate dropdown would break if you leave the page and return to the page
